Made ver.1 of peer.hook.js
